requirements.txt
pillow>=9.0.0
customtkinter>=5.0.0
